Move MappedFile class to a separate file.
