+++ cpp/CMakeLists.txt
@@ -222,6 +222,7 @@ add_library(cudf SHARED
             # src/windowed/windowed_ops.cu ... this is broken
             src/io/convert/csr/cudf_to_csr.cu
             src/io/utilities/parsing_utils.cu
+            src/io/utilities/file_utils.cpp
             src/io/convert/dlpack/cudf_dlpack.cpp
             src/io/csv/csv_reader.cu
             src/io/json/json_reader.cu
