+++ cpp/src/io/json/json_reader.cu
@@ -11,12 +11,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-
-#include <unistd.h>
-#include <sys/types.h>
-#include <sys/stat.h>
 #include <fcntl.h>
-#include <sys/mman.h>
 
 #include <thrust/device_ptr.h>
 #include <thrust/execution_policy.h>
@@ -753,28 +748,4 @@ void JsonReader::setDataTypes() {
       }
     }
   }
-}
-
-MappedFile::MappedFile(const char *path, int oflag){
-  CUDF_EXPECTS((fd_ = open(path, oflag)) != -1, "Cannot open input file");
-
-  struct stat st{};
-  if (fstat(fd_, &st) == -1 || st.st_size < 0) {
-    close(fd_);
-    CUDF_FAIL("Cannot stat input file");
-  }
-  size_ = static_cast<size_t>(st.st_size);
-}
-
-void MappedFile::map(size_t size, off_t offset){
-  CUDF_EXPECTS(size > 0, "Cannot have zero size mapping");
-  
-  map_data_ = mmap(0, size, PROT_READ, MAP_PRIVATE, fd_, offset);
-  CUDF_EXPECTS(map_data_ != MAP_FAILED, "Error mapping input file");
-  map_offset_ = offset;
-  map_size_ = size;
-}
-
-MappedFile::~MappedFile() {
-  close(fd_);
 }
